requirements.txt
numpy==1.24.2
python-telegram-bot
soundfile~=0.12.1
librosa~=0.10.0
pillow~=9.4.0
torch~=1.13.1
torchaudio~=0.13.1
speechbrain~=0.5.13
opencv-python~=4.7.0.72
onnxruntime~=1.14.1
PyYAML~=6.0
pip~=23.0.1
pytube
wheel==0.40.0
setuptools==67.6.0
packaging~=23.0
imageio~=2.26.0
scikit-image~=0.20.0
tqdm~=4.65.0
pandas~=1.5.3
torchvision~=0.14.1